refs #4 remaining refactors after update
