+++ app/models/latest_timestamp.py
@@ -7,4 +7,3 @@ class LatestTimestamp(Base):
     __tablename__ = 'latest_timestamp'
     id = Column(Integer, primary_key=True)
     timestamp = Column(BigInteger, nullable=False)
-
