+++ app/models/meta.py
@@ -1,4 +1,3 @@
-# app/models/meta.py
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import scoped_session, sessionmaker
