+++ app/services/anki_service.py
@@ -1,4 +1,5 @@
 import requests
+
 from app.app import logger
