+++ app/services/ebook_db_sync_service.py
@@ -1,9 +1,9 @@
 from sqlalchemy.exc import SQLAlchemyError
 
 from app.app import logger
-from app.models import DBSession
 from app.models.book_info import BookInfo
 from app.models.lookup import Lookup
+from app.models.meta import DBSession
 from app.models.word import Word
 
 
@@ -33,20 +33,20 @@ class EbookDBSyncService:
         logger.debug(f"Processing row: {row}")
 
         try:
-            word = Word.query.filter_by(word=word_text).first()
+            word = DBSession.query(Word).filter_by(word=word_text).first()
             if not word:
                 word = Word(word=word_text)
                 DBSession.add(word)
                 DBSession.flush()  # Generates the ID for further use
 
-            book_info = BookInfo.query.filter_by(title=book_title, authors=book_authors).first()
+            book_info = DBSession.query(BookInfo).filter_by(title=book_title, authors=book_authors).first()
             if not book_info:
                 book_info = BookInfo(title=book_title, authors=book_authors)
                 DBSession.add(book_info)
                 DBSession.flush()  # Generates the ID for further use
 
             usage_timestamp = timestamp
-            lookup = Lookup.query.filter_by(word_id=word.id, book_id=book_info.id, usage=usage_text).first()
+            lookup = DBSession.query(Lookup).filter_by(word_id=word.id, book_id=book_info.id, usage=usage_text).first()
             if not lookup:
                 lookup = Lookup(word_id=word.id, book_id=book_info.id, usage=usage_text, timestamp=usage_timestamp)
                 DBSession.add(lookup)
